refactor: move generate meta util to shared lib
