+++ apps/cms/tsconfig.json
@@ -16,6 +16,7 @@
     { "path": "../../packages/ui" },
     { "path": "../../packages/themes/base" },
     { "path": "../../packages/email" },
-    { "path": "../../packages/shared-utils" }
+    { "path": "../../packages/shared-utils" },
+    { "path": "../../packages/lib" }
   ]
 }
