+++ packages/lib/src/index.ts
@@ -1,3 +1,5 @@
 export { SHOP_NAME_RE, validateShopName } from "./validateShopName";
 export { checkShopExists } from "./checkShopExists.server";
 export { applyFriendlyZodMessages, friendlyErrorMap } from "./zodErrorMap";
+export { generateMeta } from "./generateMeta";
+export type { ProductData, GeneratedMeta } from "./generateMeta";
